agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.23.0",
    "ag-ui-adk @ git+https://github.com/ag-ui-protocol/ag-ui.git#subdirectory=integrations/adk-middleware/python",
    "uvicorn[standard]>=0.34.0",
]
